Move waste chute dimensions to opentrons.protocols.geometry.
